rag-backend/pyproject.toml
[project]
name = "rag-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.118.0",
    "langchain>=0.2.14",
    "langchain-community>=0.2.12",
    "langchain-chroma>=0.1.2",
    "langchain-openai>=0.2.2",
    "langchain-ollama>=0.1.1",
    "python-dotenv>=1.0.1",
    "pypdf>=4.2.0",
    "bs4>=0.0.2",
    "python-multipart>=0.0.20",
]
